LayoutTests/imported/w3c/web-platform-tests/css/css-tables/collapsed-scroll-overflow-expected.txt

header	body body body body body body body body body body body

FAIL collapsed-scroll-overflow assert_equals: The collapsed border does not exceed the scroll container. expected 500 but got 540
